change id type in EmployeeDTO method to Long

+++ src/main/java/com/konrad/employeemanagement/employee/EmployeeController.java
@@ -21,7 +21,7 @@ public class EmployeeController {
     @GetMapping("/{id}")
     @ResponseBody
     @ResponseStatus(HttpStatus.OK)
-    public EmployeeDTO getAllEmployees(@PathVariable("id") int id) {
+    public EmployeeDTO getAllEmployees(@PathVariable("id") Long id) {
         return employeeService.getEmployeeById(id);
     }
